refactor(dex): merge swap and transfer-out traits into dex::transport
Give the DEX-transport seam a single home ahead of routing swaps over
remote_lease (#739): ExactAmountIn and TransferOut/TransferOutFactory
now live side by side, and the task/spec vocabulary (SwapTask, Account,
SwapCoins, slippage) is usable without the `impl` feature so transport
implementors build against one module without pulling in the state
machine.
