+++ protocol/packages/dex/src/impl_/swap_exact_in/encode_req.rs
@@ -6,7 +6,7 @@ use sdk::cosmwasm_std::QuerierWrapper;
 
 use crate::{
     SlippageCalculator, SwapTask as SwapTaskT, WithCalculator, error::Result, impl_::trx::SwapTrx,
-    swap::ExactAmountIn,
+    transport::ExactAmountIn,
 };
 
 pub struct EncodeRequest<'spec, 'querier, SwapTask, SwapClient> {
