+++ protocol/packages/dex/src/impl_/swap_exact_in/mod.rs
@@ -27,7 +27,7 @@ use crate::{
         response::{self, ContinueResult, Handler, Result as HandlerResult},
         timeout,
     },
-    swap::ExactAmountIn,
+    transport::ExactAmountIn,
 };
 
 #[cfg(feature = "migration")]
