+++ protocol/packages/dex/src/lib.rs
@@ -1,13 +1,20 @@
 #[cfg(feature = "impl")]
 pub use self::error::Error;
-#[cfg(feature = "impl")]
+
 pub use self::{
     account::Account,
     anomaly::{Handler as AnomalyHandler, Treatment as AnomalyTreatment},
     coins_in::SwapCoins,
     connect::{Connectable, ConnectionParams, Ics20Channel},
-    enterable::Enterable,
     error::Result as DexResult,
+    slippage::{Calculator as SlippageCalculator, WithCalculator},
+    swap_task::{CoinsNb, SwapOutputTask, SwapTask, WithOutputTask},
+    time_alarm::TimeAlarm,
+};
+
+#[cfg(feature = "impl")]
+pub use self::{
+    enterable::Enterable,
     impl_::{
         AcceptAnyNonZeroSwap, AcceptUpToMaxSlippage, MaxSlippage, StartLocalLocalState,
         StartLocalRemoteState, StartTransferInState, StateLocalOut, StateRemoteOut, TransferOut,
@@ -15,26 +22,18 @@ pub use self::{
     },
     resp_delivery::ForwardToInner,
     response::{ContinueResult, Handler, Response, Result},
-    slippage::{Calculator as SlippageCalculator, WithCalculator},
     state::{Contract, ContractInSwap, Stage},
-    swap_task::{CoinsNb, SwapOutputTask, SwapTask, WithOutputTask},
-    time_alarm::TimeAlarm,
     transport::{
         IBC_TIMEOUT, TransferOut as TransportOut, TransferOutFactory as TransportOutFactory,
     },
 };
 
-#[cfg(feature = "impl")]
 mod account;
-#[cfg(feature = "impl")]
 mod anomaly;
-#[cfg(feature = "impl")]
 mod coins_in;
-#[cfg(feature = "impl")]
 mod connect;
 #[cfg(feature = "impl")]
 mod enterable;
-#[cfg(feature = "impl")]
 mod error;
 #[cfg(feature = "impl")]
 mod impl_;
@@ -42,14 +41,12 @@ mod impl_;
 mod resp_delivery;
 #[cfg(feature = "impl")]
 mod response;
-#[cfg(feature = "impl")]
 mod slippage;
 #[cfg(feature = "impl")]
 mod state;
-pub mod swap;
-#[cfg(feature = "impl")]
+
 mod swap_task;
-#[cfg(feature = "impl")]
+
 mod time_alarm;
-#[cfg(feature = "impl")]
-mod transport;
+
+pub mod transport;
