+++ protocol/packages/dex/src/transport/transfer.rs
@@ -1,11 +1,8 @@
 use currency::Group;
-use finance::{coin::CoinDTO, duration::Duration, instant::Instant};
+use finance::{coin::CoinDTO, instant::Instant};
 use platform::batch::Batch;
 
 use crate::SwapTask;
-
-/// IBC transfer timeout — long enough for relayers to process.
-pub const IBC_TIMEOUT: Duration = Duration::from_days(1);
 
 pub trait TransferOut
 where
