+++ protocol/packages/swap/src/testing/mod.rs
@@ -44,7 +44,7 @@ where
     .expect("Expected swap-in token to be part of selected group!")
 }
 
-pub fn from_dex_symbol<G>(symbol: &str) -> dex::swap::Result<CurrencyDTO<G>>
+pub fn from_dex_symbol<G>(symbol: &str) -> dex::transport::SwapResult<CurrencyDTO<G>>
 where
     G: Group,
 {
@@ -62,7 +62,7 @@ pub(crate) fn pattern_match_else(message_name: &str) -> ! {
 pub(crate) fn validate_a_response(resp_base64: &str, exp_amount1: Amount, exp_amount2: Amount) {
     use base64::{Engine, engine::general_purpose};
 
-    use dex::swap::ExactAmountIn;
+    use dex::transport::ExactAmountIn;
     use platform::trx;
 
     use crate::Impl;
